feat: MemorialHouseUseCaseTest 구현
